move and test convert labels and convert accuracy check examples
